學員作品/Web_氣候變遷及太陽能系統建置評估_02魏天厚_05林宣安/requirements.txt
psycopg2-binary
python-dotenv
flask
gunicorn
dash
pandas
sqlalchemy
themes
dash_bootstrap_components